+++ NEWS.md
@@ -18,6 +18,7 @@ Unreleased
 * Introduce `suspenders:prerequisites` generator
 * Introduce `suspenders:ci` generator
 * Introduce `suspenders:cleanup:organize_gemfile` task
+* Introduce `suspenders:production:environment` generator
 
 20230113.0 (January, 13, 2023)